Course functionality added

+++ query/deleteUser.js
@@ -1,8 +1,17 @@
 module.exports = {
     deleteUser: (req, res) => {
-        let query = "delete from user where userID="+req.params.ID;
-        // execute query
-        db.query(query, (err, result) => {
+	let query;
+	if(req.params.type === "user")
+        {
+	 query = "delete from user where userID="+req.params.ID;
+	}
+	if(req.params.type === "course")
+        {
+         query = "delete from course where courseID="+req.params.ID;
+        }
+
+        
+	db.query(query, (err, result) => {
             if (err) {
                 console.log(err);
             }
